+++ tests/components/sr/base/bsa/BSA.cpp
@@ -1097,6 +1097,112 @@ TEST_CASE("BSA")
       }
     }
 
+    SECTION("extract uncompressed files from v104 archive with embedded names")
+    {
+      const std::filesystem::path bsa_path{"test_sr_bsa_extractFile_idx_with_embedded_names_v104.bsa"};
+      FileGuard bsa_guard{bsa_path};
+      const auto data = "BSA\0\x68\0\0\0\x24\0\0\0\x03\x01\0\0\x02\0\0\0\x03\0\0\0\x1A\0\0\0\x19\0\0\0\0\0\0\0gn\x0As@r\xBE\x0B\x01\0\0\0]\0\0\0es\x0Es\xDC\x92\x84Z\x02\0\0\0y\0\0\0\x0Bsome\\thing\0ts\x04t'\xA7\xD0\x95$\0\0\0\xA9\0\0\0\x0Fsomething\\\x65lse\0ra\x03\x62\xC2\xA6\xD0\x95\x1E\0\0\0\xCD\0\0\0oo\x03\x66\xC2\xA6\xD0\x95%\0\0\0\xEB\0\0\0test.txt\0bar.txt\0foo.txt\0\x13some\\thing\\test.txtThis is a test.\x0A\x16something\\\x65lse\\\x62\x61r.txtfoobar\x0A\x16something\\\x65lse\\\x66oo.txtfoo was here.\x0A"sv;
+      REQUIRE( writeBsa(data, bsa_path) );
+
+      BSA bsa;
+      REQUIRE( bsa.open(bsa_path.string()) );
+      REQUIRE( bsa.grabAllStructureData() );
+
+      const char eof = static_cast<char>(std::char_traits<char>::eof());
+
+      // extract test.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_test_embedded_v104.txt" };
+        FileGuard guard{path};
+        REQUIRE_FALSE( bsa.isFileCompressed(0, 0) );
+        REQUIRE( bsa.extractFile(0, 0, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "This is a test.\n" );
+      }
+
+      // extract bar.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_bar_embedded_v104.txt" };
+        FileGuard guard{path};
+        REQUIRE_FALSE( bsa.isFileCompressed(1, 0) );
+        REQUIRE( bsa.extractFile(1, 0, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "foobar\n" );
+      }
+
+      // extract foo.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_foo_embedded_v104.txt" };
+        FileGuard guard{path};
+        REQUIRE_FALSE( bsa.isFileCompressed(1, 1) );
+        REQUIRE( bsa.extractFile(1, 1, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "foo was here.\n" );
+      }
+    }
+
+    SECTION("extract compressed files from v104 archive with embedded names")
+    {
+      const std::filesystem::path bsa_path{"test_sr_bsa_extractFile_idx_compressed_with_embedded_names_v104.bsa"};
+      FileGuard bsa_guard{bsa_path};
+      const auto data = "BSA\0\x68\0\0\0\x24\0\0\0\x07\x01\0\0\x02\0\0\0\x03\0\0\0\x1A\0\0\0\x19\0\0\0\0\0\0\0gn\x0As@r\xBE\x0B\x01\0\0\0]\0\0\0es\x0Es\xDC\x92\x84Z\x02\0\0\0y\0\0\0\x0Bsome\\thing\0ts\x04t'\xA7\xD0\x95/\0\0\0\xA9\0\0\0\x0Fsomething\\\x65lse\0ra\x03\x62\xC2\xA6\xD0\x95+\0\0\0\xD8\0\0\0oo\x03\x66\xC2\xA6\xD0\x95\x32\0\0\0\x03\x01\0\0test.txt\0bar.txt\0foo.txt\0\x13some\\thing\\test.txt\x10\0\0\0x\x9C\x0B\xC9\xC8,V\0\xA2\x44\x85\x92\xD4\xE2\x12=.\0.\xC5\x05.\x0A\x16something/else/bar.txt\x07\0\0\0x\x9CK\xCB\xCFOJ,\xE2\x02\0\x0B/\x02\x84\x0A\x16something/else/foo.txt\x0E\0\0\0x\x9CK\xCB\xCFW(O,V\xC8H-J\xD5\xE3\x02\0&&\x04\xAC\x0A"sv;
+      REQUIRE( writeBsa(data, bsa_path) );
+
+      BSA bsa;
+      REQUIRE( bsa.open(bsa_path.string()) );
+      REQUIRE( bsa.grabAllStructureData() );
+
+      const char eof = static_cast<char>(std::char_traits<char>::eof());
+
+      // extract test.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_test_compressed_embedded_v104.txt" };
+        FileGuard guard{path};
+        REQUIRE( bsa.isFileCompressed(0, 0) );
+        REQUIRE( bsa.extractFile(0, 0, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "This is a test.\n" );
+      }
+
+      // extract bar.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_bar_compressed_embedded_v104.txt" };
+        FileGuard guard{path};
+        REQUIRE( bsa.isFileCompressed(1, 0) );
+        REQUIRE( bsa.extractFile(1, 0, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "foobar\n" );
+      }
+
+      // extract foo.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_foo_compressed_embedded_v104.txt" };
+        FileGuard guard{path};
+        REQUIRE( bsa.isFileCompressed(1, 1) );
+        REQUIRE( bsa.extractFile(1, 1, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "foo was here.\n" );
+      }
+    }
+
     SECTION("extract compressed files from v105 archive")
     {
       const std::filesystem::path bsa_path{"test_sr_bsa_extractFile_idx_with_compression_v105.bsa"};
@@ -1124,7 +1230,7 @@ TEST_CASE("BSA")
         std::getline(stream, content, eof);
         REQUIRE( content == "This is a test.\n" );
         #else
-        // Cannot extract lz4-compressed content with LZ4 support.
+        // Cannot extract lz4-compressed content without LZ4 support.
         REQUIRE_FALSE( bsa.extractFile(0, 0, path.string()) );
         #endif
       }
@@ -1157,7 +1263,128 @@ TEST_CASE("BSA")
         std::getline(stream, content, eof);
         REQUIRE( content == "foo was here.\n" );
         #else
-        // Cannot extract lz4-compressed content with LZ4 support.
+        // Cannot extract lz4-compressed content without LZ4 support.
+        REQUIRE_FALSE( bsa.extractFile(1, 1, path.string()) );
+        #endif
+      }
+    }
+
+    SECTION("extract uncompressed files from v105 archive with embedded names")
+    {
+      const std::filesystem::path bsa_path{"test_sr_bsa_extractFile_idx_with_embedded_names_v105.bsa"};
+      FileGuard bsa_guard{bsa_path};
+      const auto data = "BSA\0\x69\0\0\0\x24\0\0\0\x03\x01\0\0\x02\0\0\0\x03\0\0\0\x1A\0\0\0\x19\0\0\0\0\0\0\0gn\x0As@r\xBE\x0B\x01\0\0\0\0\0\0\0m\0\0\0\0\0\0\0es\x0Es\xDC\x92\x84Z\x02\0\0\0\0\0\0\0\x89\0\0\0\0\0\0\0\x0Bsome\\thing\0ts\x04t'\xA7\xD0\x95$\0\0\0\xB9\0\0\0\x0Fsomething\\\x65lse\0ra\x03\x62\xC2\xA6\xD0\x95\x1E\0\0\0\xDD\0\0\0oo\x03\x66\xC2\xA6\xD0\x95%\0\0\0\xFB\0\0\0test.txt\0bar.txt\0foo.txt\0\x13some\\thing\\test.txtThis is a test.\x0A\x16something\\\x65lse\\\x62\x61r.txtfoobar\x0A\x16something\\\x65lse\\\x66oo.txtfoo was here.\x0A"sv;
+      REQUIRE( writeBsa(data, bsa_path) );
+
+      BSA bsa;
+      REQUIRE( bsa.open(bsa_path.string()) );
+      REQUIRE( bsa.grabAllStructureData() );
+
+      const char eof = static_cast<char>(std::char_traits<char>::eof());
+
+      // extract test.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_test_embedded_v105.txt" };
+        FileGuard guard{path};
+        REQUIRE_FALSE( bsa.isFileCompressed(0, 0) );
+        REQUIRE( bsa.extractFile(0, 0, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "This is a test.\n" );
+      }
+
+      // extract bar.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_bar_embedded_v105.txt" };
+        FileGuard guard{path};
+        REQUIRE_FALSE( bsa.isFileCompressed(1, 0) );
+        REQUIRE( bsa.extractFile(1, 0, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "foobar\n" );
+      }
+
+      // extract foo.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_foo_embedded_v105.txt" };
+        FileGuard guard{path};
+        REQUIRE_FALSE( bsa.isFileCompressed(1, 1) );
+        REQUIRE( bsa.extractFile(1, 1, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "foo was here.\n" );
+      }
+    }
+
+    SECTION("extract compressed files from v105 archive with embedded file names")
+    {
+      const std::filesystem::path bsa_path{"test_sr_bsa_extractFile_idx_compressed_with_embedded_names_v105.bsa"};
+      FileGuard bsa_guard{bsa_path};
+      const auto data = "BSA\0\x69\0\0\0\x24\0\0\0\x07\x01\0\0\x02\0\0\0\x03\0\0\0\x1A\0\0\0\x19\0\0\0\0\0\0\0gn\x0As@r\xBE\x0B\x01\0\0\0\0\0\0\0m\0\0\0\0\0\0\0es\x0Es\xDC\x92\x84Z\x02\0\0\0\0\0\0\0\x89\0\0\0\0\0\0\0\x0Bsome\\thing\0ts\x04t'\xA7\xD0\x95\x37\0\0\0\xB9\0\0\0\x0Fsomething\\\x65lse\0ra\x03\x62\xC2\xA6\xD0\x95\x31\0\0\0\xF0\0\0\0oo\x03\x66\xC2\xA6\xD0\x95\x38\0\0\0!\x01\0\0test.txt\0bar.txt\0foo.txt\0\x13some/thing/test.txt\x10\0\0\0\x04\"M\x18`@\x82\x10\0\0\x80This is a test.\x0A\0\0\0\0\x16something/else/bar.txt\x07\0\0\0\x04\"M\x18`@\x82\x07\0\0\x80\x66oobar\x0A\0\0\0\0\x16something/else/foo.txt\x0E\0\0\0\x04\"M\x18`@\x82\x0E\0\0\x80\x66oo was here.\x0A\0\0\0\0"sv;
+      REQUIRE( writeBsa(data, bsa_path) );
+
+      BSA bsa;
+      REQUIRE( bsa.open(bsa_path.string()) );
+      REQUIRE( bsa.grabAllStructureData() );
+
+      const char eof = static_cast<char>(std::char_traits<char>::eof());
+
+      // extract test.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_test_embedded_compressed_v105.txt" };
+        FileGuard guard{path};
+        REQUIRE( bsa.isFileCompressed(0, 0) );
+        #if !defined(MWTP_NO_LZ4)
+        REQUIRE( bsa.extractFile(0, 0, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "This is a test.\n" );
+        #else
+        // Cannot extract lz4-compressed content without LZ4 support.
+        REQUIRE_FALSE( bsa.extractFile(0, 0, path.string()) );
+        #endif
+      }
+
+      // extract bar.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_bar_embedded_compressed_v105.txt" };
+        FileGuard guard{path};
+        REQUIRE( bsa.isFileCompressed(1, 0) );
+        #if !defined(MWTP_NO_LZ4)
+        REQUIRE( bsa.extractFile(1, 0, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "foobar\n" );
+        #else
+        // Cannot extract lz4-compressed content without LZ4 support.
+        REQUIRE_FALSE( bsa.extractFile(1, 0, path.string()) );
+        #endif
+      }
+
+      // extract foo.txt
+      {
+        const auto path { std::filesystem::temp_directory_path() / "extract_foo_embedded_compressed_v105.txt" };
+        FileGuard guard{path};
+        REQUIRE( bsa.isFileCompressed(1, 1) );
+        #if !defined(MWTP_NO_LZ4)
+        REQUIRE( bsa.extractFile(1, 1, path.string()) );
+
+        std::ifstream stream(path, std::ios::in | std::ios::binary);
+        std::string content;
+        std::getline(stream, content, eof);
+        REQUIRE( content == "foo was here.\n" );
+        #else
+        // Cannot extract lz4-compressed content without LZ4 support.
         REQUIRE_FALSE( bsa.extractFile(1, 1, path.string()) );
         #endif
       }
@@ -1317,7 +1544,7 @@ TEST_CASE("BSA")
         std::getline(stream, content, eof);
         REQUIRE( content == "This is a test.\n" );
         #else
-        // Cannot extract lz4-compressed content with LZ4 support.
+        // Cannot extract lz4-compressed content without LZ4 support.
         REQUIRE_FALSE( bsa.extractFile("some\\thing\\test.txt", path.string()) );
         #endif
       }
@@ -1335,7 +1562,7 @@ TEST_CASE("BSA")
         std::getline(stream, content, eof);
         REQUIRE( content == "foo was here.\n" );
         #else
-        // Cannot extract lz4-compressed content with LZ4 support.
+        // Cannot extract lz4-compressed content without LZ4 support.
         REQUIRE_FALSE( bsa.extractFile("something\\else\\foo.txt", path.string()) );
         #endif
       }